tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/037.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: DOM movement with appendChild, inline</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>
	<div></div>
	<script>log('inline script #1');
		var script = testlib.addScript('log(\'head script #1\')', {}, document.getElementsByTagName('head')[0], true);
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	try{
		document.body.appendChild(script);
	}catch(e){ log('ERROR - tested functionality not supported'); }
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'head script #1', 'end script #1', 'inline script #2' ]);
		t.done();
}
	onload = t.step_func(function(){
		setTimeout(t.step_func(test), 200);
	});
	</script>

</body></html>
